Fix log4net tests to account for new process span to check for Android (#7104)
## Summary of changes

Fixes the log4net tests for v3.1.0 which caused us to create a new
process span when log4net is checking if the OS is Android.

## Reason for change

Tests were broken due to changes in
https://github.com/apache/logging-log4net/releases/tag/rel%2F3.1.0 with
how they detect Android.

## Implementation details

Just ignored the process span - the call is only made on non-Windows for
log4net and only in v3.1.0 and above.

## Test coverage

Updated tests to exclude the process span.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: andrewlock <18755388+andrewlock@users.noreply.github.com>
Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>