Remove flaky datapipeline tests (#7124)
## Summary of changes

Remove datapipeline tests which are causing a lot of flakiness,
particularly with named pipes

## Reason for change

These tests are failing a _lot_ due to duplicate spans in the payload.
It needs investigating, and this PR should be reverted once the issue
has been fixed.

## Implementation details

Delete the data pipeline tests. It's not easy to explicitly skip them,
because of their usage in theory tests rather than facts

## Test coverage

Less now. This must be reverted before we enable datapipeline usage