Improve CharSlice allocations (#7127)
## Summary of changes

This PR reduces the number of allocation of the CharSlice type

## Reason for change

The current code allocates in both Managed and Unmanaged heap for each
instance, we can skip the managed allocation.

## Implementation details

## Test coverage

The struct is used in several parts of the code, so current tests covers
the change.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->