Add bailout scenario to execution time tests (#7136)
## Summary of changes

Add "bailout" scenario to execution time tests

## Reason for change

We want to make sure that we're not adding any significant overhead when
we bail-out of instrumentation (e.g. because the app is on the exclude
list)

## Implementation details

Added a new scenario to the execution-time benchmarks in which we set
`DD_PROFILER_EXCLUDE_PROCESSES`, to emulate a standard bailout scenario.
We can then compare the `process.time_to_start_ms` values between
baseline (no instrumentation) and bailout, which gives the time elapsed
between timeit `cmd.Start()` to the `Starthook` mehod.

## Test coverage

Overall, the benchmarks show that bailout adds ~1-2ms for both .NET Core
and .NET Framework, on top of a ~43ms start time for a .NET Core app

## Other details

I originally added the test to the HttpMessageHandler test, but that
test already takes a little longer, so moved it to the FakeDbCommand
instead. The results for startup time are essentially iddentical in both
cases anyway