Replace "manual" retries with buit-in flaky retry (#7132)
## Summary of changes

- Replace the manual retries we are doing in tests with the `[Flaky]`
attribute
- Add `SkipOn.AllExcept()` helpers
- Fixed a bug in the `[FlakyAttribute]` metric submission

## Reason for change

- Reduces duplication
- Easier to grep for
- Same net results

## Implementation details

Searched for `ReportRetry()` and then a lot of copy, paste, and delete.
I split up the `AgentMalfunctionTests` to avoid marking the non-flaky
tests (UDS/TCP) as flaky.

Adding the `SkipOn.AllExcept()` makes it easier to say "only run this on
Windows" for example. Previously you had to exclude all the other
platforms, or do the checks yourself.

Finally, fixed a bug in the metric submission for the `[Flaky]`
attribute, which was sending invalid JSON

## Test coverage

Should be the same.