Wait for config sent in Metastruct tests (#7130)
## Summary of changes

Some security tests seem to be flaky. Metastruct data seems to be
missing sometimes.
A likely cause is that the agent config (sent through http) is not yet
received when the test hits the vulnerability, so metaStruct is not yet
set as supported.

```
    internal bool IsMetaStructSupported()
    {
        if (_discoveryService is null)
        {
            _discoveryService = Tracer.Instance.TracerManager.DiscoveryService;
            SubscribeToDiscoveryService(_discoveryService);
        }

        return _spanMetaStructs;
    }
```

In this PR, when metastruct is used in an integration test, we wait
until the configuration is sent by the mock agent before actually making
the request.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->