Don't write native loader logs when bailing out in Windows SSI (#7147)
## Summary of changes

- When we're running in Windows SSI, buffer logs for native loader until
we know whether we're bailing out. If we bail out, drop the logs,
otherwise, write them and enable-auto flushing
- Refactor `InstrumentationTests`

## Reason for change

In Windows SSI, we inject into every process by default, and then bail
out. Today, this results in creating a native loader file for every
injection, which causes a lot of noise in the log files. With the
updated implementation, we buffer those logs until we know we're
continuing with instrumentation instead, so we don't create loads of log
files.

## Implementation details

- Use spdlog's "backtrace" support, that allows buffering logs without
writing them.
- Introduce an `enable_buffering()` method to the `TLoggerPolicy`
abstraction, to know whether logs should be buffered initially, or just
auto-flushed
- This returns `false` for Tracer and Profiler, because by the time
they're running, we know we want to instrument.
- For the native loader, we look for the SSI variable. If it's set, we
check if buffering is explicitly enabled/disabled (default is enabled).
This ensures that if we have a crash (for example) that we have a way of
forcing logs to be written.
- If buffering is **not enabled**, the existing behavior (pre this PR)
is used.
- If buffering is **enabled**:
  - We disable automatic flushing
  - Set the log level to be "don't write any logs"
  - Enable back trace
- Use a `lazy_rolling_file_sink`. This is necessary because the built-in
`rolling_file_sink` creates and opens the target file as soon as it's
created. The updated version delays this until we write to the sink
- Later, if we determine that we _do_ want to instrument the app, we
call `FlushAndDisableBuffering` to restore the original behaviour

## Test coverage

I refactored the existing instrumentation tests 
- Don't not rely on `dotnet run`, because is a bit messy because we
instrument it _and_ the target app (we should look into this seperately)
- Reduce duplication
- Add expanded assertions that we don't have native tracer or profiler
in "normal" bailout scenarios
- Add additional tests and assertions for the new behavaiour in this PR
- Updated execution benchmarks scenario to emulate bailout in SSI (i.e.
no logs scenario)