Remove all-green jobs (#7173)
## Summary of changes

Remove all-green jobs

## Reason for change

Right now, regarding the current state of CI, it's not a good fit : as
many jobs are retried during PR, and as github action has no way to know
that a job has been retried, the all-green jobs must be manually
retried., causing friction.

In practice, the all-green jobs is not retried, making the all-green job
the main source of "red" pull request. Which is the opposite of what
it's meant.

The guild will discuss about implementing the same pattern, but inside
the azure pipeline (so it'll be automatically retried).

 
## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->