Fix github token usage (#7174)
## Summary of changes

Fix the github status updates from Azure DevOps

## Reason for change

In #7091 we switched to how we were obtaining the token. We missed one
place it's needed.

## Implementation details

Obtain a token instead of trying to use an ambient token

## Test coverage

This is the test

## Other details

Oops 🤦‍♂️ 

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->