Add retry to smoke crash tracking tests (#7172)
## Summary of changes

Retry the crash-test smoke tests if the sample fails to exit after crash
analysis

## Reason for change

We have an intermittent hang _after_ performing the crash analysis in
crash tests. We don't have any good leads in terms of investigating the
issue, so to avoid impacting devex, just retry the job.

## Implementation details

We already have a timeout, so just add a retry. This will retry when we
explicitly fail too, but that's _probably_ not a big concern. If it is,
we will have to [manage the timeout in bash
itself](https://stackoverflow.com/questions/687948/how-to-run-a-command-with-a-timeout-so-that-it-is-killed-if-it-exceeds-the-timeo)
but I don't think it's worth the hassle

## Test coverage

I'll retry it a few times and see if we catch a failure 😅 

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->