Make sure we can override the `DD_PROFILING_ENABLED` value in smoke tests (#7184)
## Summary of changes

- Make sure we can set `DD_PROFILING_ENABLED=0` for the crashtracking
tests

## Reason for change

There's a known interaction with the crash-tracking tests and the
profiler which can cause the crash tracker to hang. We have set
`DD_PROFILING_ENABLED=0` as a workaround for this for some time, but we
were still seeing issues with `dd-dotnet`. It turns out that we were
accidentally forcing it back on in those tests, so this should fix it

## Implementation details

Instead of explicitly calling `--set-env` in the dd-dotnet entrypoint,
set the variable _globally_, so that it can be disabled as required in
the crash tracking test here:


https://github.com/DataDog/dd-trace-dotnet/blob/283c79a11b324511db1dc30cb16b4146d3e93fbd/.azure-pipelines/steps/run-snapshot-test.yml#L157

## Test coverage

This is the test - hopefully all the installer tests pass!