[fleet installer] Remove the global instrumentation commands (#7182)
## Summary of changes

- Removes the global instrumentation command
- Moves the global instrumentation logic into IIS command (currently
unused)
- Removes non-applicable smoke tests

## Reason for change

We are changing the approach to shipping the command, and don't want to
expose an additional command, as discussed with @bmermet

## Implementation details

- Move the logic previously in the "global" commands into the IIS
commands
- Refactor, so it's easy to switch to the alternative
- Remove the smoke tests that test with the "global" commands

## Test coverage

As long as the existing smoke tests pass, this is fine.

## Other details

We jumped the gun a bit merging those commands, so are backing them out.
See https://github.com/DataDog/dd-trace-dotnet/pull/7183 for an
alternative approach we're looking at.