Try to fix flaky IIS fleet-installer tests (#7177)
## Summary of changes

Add workaround for flaky ASM tags on fleet installer IIS tests 

## Reason for change

The IIS fleet-installer tests are kind of gnarly and slow, and we have
been seeing some flay related to the presence (or lack of) the
`_dd.appsec.waf.version` tag on spans where we expect them. We tried to
solve the flakiness at the source in
https://github.com/DataDog/dd-trace-dotnet/pull/7121, but we're still
seeing flakiness, so for now just ignoring the tag in the snapshot
comparison.

## Implementation details

Unfortunately, this is a bit more tricky than we'd like because we
_only_ want to ignore it for IIS atm. That means we have to create a
separate docker target, and add the ability to specify the target to
run.

## Test coverage

Covered by existing tests - if they pass, that's good enough for now

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->