Revert "Fix test InitializesTracerWhenTracingIsDisabled Flakiness (#7163) (#7191)
## Summary of changes

This reverts commit 814c3b9385fa0405fb630d54adb403fc776118d4.

## Reason for change

It caused massive flakiness in the Debugger tests on Windows. I think it
exacerbates an existing flakiness present in the `LogEntryWatcher`. It
appears the `NewLogFileCreated` even is often not triggered. Will
re-implement and investigate in a separate PR, this is just to unblock
things.

## Implementation details

Revert the previous PR

## Test coverage

I'll do a dedicated run of debugger tests to confirm this resolves the
issue

## Other details

The debugger tests and the tracer test _were_ flaky, so it seems likely
that this was exacerbating existing flake rather than introducing it