Fix potential flake in crash-tracking tests for chiseled containers (#7190)
## Summary of changes

Ensure we don't run crash-tests with the profiler enabled for chiseled
containers

## Reason for change

We have a known flake interaction between the CP and crash tracking. We
merged a fix for most cases in
https://github.com/DataDog/dd-trace-dotnet/pull/7184, this handles the
chiseled container scenario

## Implementation details

We can't use `ENV` in chiseled containers, so we can't move the
`DD_PROFILING_ENABLED` call to the ambient container (and override it
for the crash tracking tests). So instead, we remove it entirely, and
then explicitly enable/disable for _both_ the crash-tracking and normal
tests intead.

## Test coverage

This is the test, as long as everything passes, we're ok.