Try to fix flaky tests that use `FileSystemWatcher` (#7189)
## Summary of changes

- Re-implements #7163 to get visibility on `LogEntryWatcher` behaviour
- Ensure tests that use `LogEntryWatcher` write logs to separate
locations
- Increase notification buffer size of `FileSystemWatcher`
- Add logging if/when buffer overflows

## Reason for change

#7163 was attempting to improve our visibility of the behaviour of the
`LogEntryWatcher` (which uses a `FileSystemWatcher`) to try to
understand flake we were seeing. However, that PR _massively_ increased
the flake in the Debugger tests on Windows (to the extent that we had no
successful runs any more). So we reverted that PR in #7191.

This PR reinstates the original PR, but adds additional mitigations, to
attempt to avoid the issue repeating.

## Implementation details

The first commit is a simple revert of the "revert PR" #7191 (no need to
review it closely).

The mitigations to prevent repeat issues are:
- Don't write all logs to the same folder. Theoretically I thought that
this _shouldn't_ make a difference, as we already split log folders by
class, and these aren't run concurrently, but on its own, this resolved
the issue, so it clearly does have an impact 🤷‍♂️
- Increase the size of the notification buffer. A notification buffer
overflow is the most likely cause of the issue, especially as the error
happened the most on debug runs, when the most file-change notifications
will be sent.
- Add a notification if the buffer overflows, so we know _why_ we're
failing.

## Test coverage

Ran tests locally - previously I saw flake, afterwards I didn't, and the
same goes for CI.

## Other details

I updated everywhere I found that was using the `LogEntryWatcher`, even
those that haven't seen flake so far, just in case.