[Dynamic Instrumentaiton] Marking instrumentation as failure if IsReEntryField is absent (#7188)
## Summary of changes
In DI rewriting, we inject a field in all state machines to track the
state between re-entrances of async flow. If we fail to inject that
field on module load, we mark the method as not eligible for
instrumentation of method probes.
Line Probes, on the other hand, we don't need to track re-entrances thus
they stay intact if we fail to inject that failed. In the current state
of the instrumentation, we returned S_OK from the method probe
instrumentation even though we failed to apply to to let line probes
proceed. It lead to the situation where we fixed Exception Handler
Clauses even though we didn't have to, leading to
`InvalidProgramException` as a result.