[Exception Replay] Groundwork to support Test Optimization (#7085)
## Summary of changes
CI Test Optimization starts to utilize Exception Replay to capture debug
information for failing tests. If a test fails in CI, the test will run
multiple times - with Exception Replay enabled - and that will surface
the variable values of the methods participating in the stack trace of
the exception that has failed the test. This is only true for tests that
fail due to _exception_ and not logical failures.

## Reason for change
Added a few helpers for Test Optimization to use Exception Replay
optimally:
1. Avoid reverting exception cases. Exception Replay by default captures
only 1 occurrence of an exception and then avoid capturing it for
certain amount of time. For Test Optimization we want to capture several
occurrences at the same time. Rate limit is now able to be 0 - meaning,
skip reverting & re-instrumenting.
2. Test Optimization needs to be notified when Exception Replay started
to track an exception - meaning, instrumentation is underway and
capturing is possible - so it will trigger the failing test repeatedly.