Adding Throughput Test for Kafka & DSM (#7133)
## Summary of changes
We want to enable Data Streams Monitoring by default. To ensure that the
DSM code path has a minimal impact on customers, we've added throughput
testing to ensure that the overhead of DSM is acceptable in various
scenarios.

This PR adds throughput testing for publishing and consuming N Kafka
messages. We compare the runtime of DSM with APM and APM without DSM
enabled to ensure we have an acceptable performance overhead.

If the overhead of DSM is >= the given threshold (currently 4%), we fail
the CI check.

See
[here](https://gitlab.ddbuild.io/DataDog/apm-reliability/dd-trace-dotnet/-/jobs/1017470789)
for a successful run.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->