fix: Handle both SQS StringValue and SNS BinaryValue in context extraction (#7205)
## Summary of changes
SNS -> SQS context propagation was broken, as we encode the value of the
`_datadog` header into base64. On the consume side, we assume it is
written as a string, which is only true if we send a message directly
through SQS.

The fix is to decode the `_datadog` value based on the type, just as in
[this](https://github.com/DataDog/datadog-lambda-js/pull/269) PR.

This soles issues where the DSM map doesn't show links between SNS and
SQS, as the checkpoint is still created on the consumer, just without a
link to the parent.

## Reason for change

## Implementation details

## Test coverage
I tested this via system-tests, with a description of my test
methodology in [this
PR](https://github.com/DataDog/system-tests/pull/4897).
The SNS system test above pass when I ran locally with this tracer
build.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->