Only enable Schema extraction when DSM explicitly enabled (#7175)
## Summary of changes
DSM enabled by default will soon happen, but we don't want the schema
extraction logic to be enabled by default due to performance concerns
and the addition of potentially large span tags. This should only be
enabled when DSM is explicitly enabled, so that this behaviour works as
is for existing DSM clients.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->