Only download samples for the framework that we're testing (#7227)
## Summary of changes

For integration tests, only download the standalone samples for the
framework that we're testing

## Reason for change

Currently we
- Build the "standalone" samples for all TFMs (which is the easiest way
to do it)
- Build the "multi-version" samples separately for each TFM

In integration tests, we currently pull down _all_ the standalone
samples. For platforms that need it, we also pull down the multi-version
samples for the TFM we're testing.

The net result is that we pull down a bunch of extra built samples that
we don't need. This is causing problems as we're actually running out of
drive space in the .NET 10 upgrade branch 😅

## Implementation details

Continue to build the samples in a "single shot" as this is optimal in
terms of VM usage, and upload as a single blob. This is unchanged, and
it looks like this today:

<img width="220" height="469" alt="image"
src="https://github.com/user-attachments/assets/eb39a343-1b21-4b26-a39b-c422919bf87b"
/>

On the download side, only download the artifacts for the TFM we're
currently testing. We always split by TFM, and the pattern of artifacts
is well known, so this works fine.

## Test coverage

This is the test, as long as the build passes we're good. Appears to
reduce the downloaded size from ~8GB to ~300-600MB.

## Other details

Blocker for https://github.com/DataDog/dd-trace-dotnet/pull/7170.

Other options considered:
- Split the `build_samples.standalone` job by TFM. This is doable, but
ends up using more VM time overall due to the individual VM overhead.
Currently we're very well matched.
- Combine the single TFM build into the multi-sample build. Would use
one less VM, but would increase the length of the stage overall. Also
would use more drive space for cases where we _don't_ need the
multi-sample build (ie. `integration_tests_windows`, which is where we
have the biggest space issues), but it's possible.